EditorShellのDocking実装を物理分離
